+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v3.127.0
+* Changed behaviour of `table.Client` and `query.Client` internal session pool. When `ydb.WithPreferredNodeID` is set and there is no idle session on preferred node, the pool closes most idle session to create a new one on the preferred node
+* Fixed a bug when sometimes in cases of context cancellation returner error was not `errors.Is(err, context.Canceled)`
+
+## v3.126.5
 * Invoke `OnWriterReceiveResult` trace callback when topic writer receives ack from server (enables client-side logging for sync write diagnostics)
 * Added `ydb.WithCommitTxContext(ctx)` to request commit along with query execution in database/sql transactions